api/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
supabase==2.5.0
openai==1.35.3
pytest==8.2.2
youtube-transcript-api==0.6.2
firecrawl-py==1.0.0
